share InitializeFolders implementation for OS
